Update README.md
Welcome to LADYS, the meme coin of Milady NFT collection. LADYS is the tokenisation of the fully memetically optimized white pill.
LADYS is not another meme coin. LADYS is a self-organised meme coin. LADYS is the drip currency. LADYS is the points for karma, charm and beauty. LADYS is the accumulation of clout.Milady meme coin is LADYS, and LADYS is the meme coin for Milady.